+++ CHANGELOG.md
@@ -12,8 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Objective` now has a `metadata` attribute as well as a `description` property
 - `Target` and `Parameter` now have a `metadata` attribute as well as `description` and
   `unit` properties
-- `FPSRecommender` now optionally uses the fpsample library with fallback
-  to internal implementation
+- `FPSRecommender` now optionally uses the fpsample library (if installed) with fallback
+  to internal implementation. The use of `fpsample` can be deactivated by setting the
+  environment variable `BAYBE_ENFORCE_FALLBACK_FPS_IMPLEMENTATION` to a truthy value
 - `farthest_point_sampling` now also supports a collection of integers for 
   `initialization`, using them for pre-selecting points